set up useEffect for cities
